Add files via upload
We have the dataset of the outcome of a  bank marketing campaign. The product is Term deposit. Using Machine learning algorithms, we have to build a model to predict the outcome of future campaigns i.e. whether the customer will subscribe to the term deposit or not.
The solution contains the step-by-step approach taken to build different models, the theory behind every step, model comparison, model validation parameters, and selection of the best fit model. This project can explain all the basic and advanced machine learning concepts and their implementation in the industry.
Dataset from Kaggle